ASoC: fsl: fsl_sai: fix random channel mapping when using continuous clocks [STREAMEIGHT-4553]
When using continuous clocks, it's been discovered that the channel
mapping is unstable.

For 8 channels, channel 1 data can be mapped to either channel 1 or 3,
channel 2 data to either channel 5 or 7,
channel 3 data to either channel 2 or 4,
channel 4 data to either channel 6 or 8,
channel 5 data to either channel 1 or 3,
channel 6 data to either channel 5 or 7,
channel 7 data to either channel 2 or 4,
channel 8 data to either channel 6 or 8.

This behavior isn't reproducible on non-continuous clocks mode (though
the mapping is still incorrect).

In the usual code path in SNDRV_PCM_TRIGGER_STOP,
SNDRV_PCM_TRIGGER_SUSPEND and SNDRV_PCM_TRIGGER_PAUSE_PUSH FIFOs are
reset. The issue is that resetting FIFOs requires the disabling of
transmitter and receiver which disable the clocks. One cannot do that
when in continuous clocks mode. Therefore the code path was dropped.

Since FIFOs resets seem to be required in order to have fixed channel
mapping and that one cannot reset the FIFOs without disabling
transmitter and receiver, the process is done as late as possible to
avoid clock glitches as much as possible.

In that effort, Bit Clock Enable bit is also set when transmitter and
receiver are disabled in an attempt to lower the possibility of having
clock glitches.

The software resetting of transmitter and receiver has voluntarily been
left out when copying code path from SNDRV_PCM_TRIGGER_STOP since it
seemed overkill and it looks like it works just fine without it.

As checked on a logic analyzer, BCLK and LRCLK/WCLK do not seem to be
impacted by this change.

Fixes: 9f6f1fe7624b "ASoC: fsl_sai: add handling of SND_SOC_DAIFMT_CONT"

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>